Update the outbound shipment order document
We updated the outbound shipment order document since the Spree::ShippingMethod
model already has the carrier and service_level field that can be used to define
the carrier_name and service_level needed to QL.

We also added the field priority to the Spree::ShippingMethod model with
STANDARD as default.
With these changes the shipping_attributes proc is no longer needed.