Fix quantityOrdered and quantityToShip attributes on the SO document
If there are only 5 stock items for the ordered product, and the 
requested quantity is 10, the order will be created with 2 shipments.
Each shipment has one line item with quantity set to 10.
The problem is that each shipment is an order for QL so we should take 
the inventory units count for each shipment instead of the line item 
quantity otherwise QL will receive double the items to be shipped.

Add the regression test.